feat: Add environment variables configuration
The code changes in this commit add PostgreSQL, email, and JWT environment variables to the `.env.example` file. These variables are necessary for configuring the database connection, email service, and JSON Web Token functionality in the application. Additionally, the scripts `dev` were added to the `package.json` to execute .env.local for development and `prod` to execute .env.production for production.