feat: Fix nombre field in getByResProd function
The `getByResProd` function in the `product.controller.js` file was retrieving the `nombre` parameter from the request body instead of the query parameters. This commit fixes the issue by updating the code to retrieve the `nombre` parameter from the query parameters. This ensures the correct functionality of selecting all products of a restaurant by name.