feat: Improve "getById" handler to get user data
This commit modifies the "getById" controller in the "user_data.controller.js" file to improve fetching user data. The option to obtain additional information is added depending on the type of user (customer, administrator or restaurant). This improves the functionality of the endpoint by providing specific data for each type of user.